move validate tests to test_dbs.py
