+++ q2_moshpit/eggnog/tests/test_dbs.py
@@ -12,7 +12,7 @@ from qiime2.core.exceptions import ValidationError
 from .._dbs import (
     fetch_eggnog_db, build_custom_diamond_db, fetch_eggnog_proteins,
     fetch_diamond_db, build_eggnog_diamond_db, fetch_ncbi_taxonomy,
-    _collect_and_compare_md5, fetch_eggnog_hmmer_db
+    _collect_and_compare_md5, fetch_eggnog_hmmer_db, _validate_taxon_id
 )
 from q2_types.feature_data import ProteinSequencesDirectoryFormat
 from q2_types.reference_db import (
@@ -35,6 +35,15 @@ class TestFetchDB(TestPluginBase):
             "--data_dir", str(eggnog_db)
         ]
         subp_run.assert_called_once_with(cmd, check=True)
+
+    def test_validate_taxon_id_invalid(self):
+        path_to_data = self.get_data_path('build_eggnog_diamond_db/')
+        with self.assertRaisesRegex(ValueError, "'0' is not valid taxon ID. "):
+            _validate_taxon_id(path_to_data, 0)
+
+    def test_validate_taxon_id_valid(self):
+        path_to_data = self.get_data_path('build_eggnog_diamond_db/')
+        _validate_taxon_id(path_to_data, 2)
 
     @patch('tempfile.TemporaryDirectory')
     @patch("q2_moshpit.eggnog._dbs._download_and_build_hmm_db")
